refactor: fix and organize imports across multiple files
